Digital Thermometer
Make your own Digital Thermometer and monitor temperature over internet from anywhere.This instructable is a basic one to start tinkering with IoT. We will be interfacing temperature sensor LM35 with NodeMCU 1.0(ESP-12E).

LM35 is a temperature sensor which can measure temperature in the range of -55°C to 150°C.It is a 3-terminal device that provides analog voltage proportional to the temperature. NodeMCU ADC can be used to measure the analog voltage from LM35 and hence calculate the temperature which is in proportion to the analog voltage.